expose, correct, unify defaults and wire up (missing) playfield/flipper physics im/export settings
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@2516 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ flasher.cpp
@@ -413,7 +413,7 @@ void Flasher::RenderSetup(RenderDevice* pd3dDevice)
    PolygonToTriangles(vvertex, vpoly, vtri);
    }
 
-   numPolys = vtri.size()/3;
+   numPolys = (int)(vtri.size()/3);
    if (numPolys == 0)
    {
       // no polys to render leave vertex buffer undefined
