+++ surface.cpp
@@ -704,7 +704,7 @@ void Surface::GenerateMesh(std::vector<Vertex3D_NoTex2> &topBuf, std::vector<Ver
       PolygonToTriangles(vvertex, vpoly, topBottomIndices);
       }
 
-      numPolys = topBottomIndices.size() / 3;
+      numPolys = (unsigned int)(topBottomIndices.size() / 3);
       if (numPolys == 0)
       {
          // no polys to render leave vertex buffer undefined
